+++ src/Cortside.AspNetCore.EntityFramework/Searches/ISearch.cs
@@ -1,4 +1,4 @@
-namespace Cortside.AspNetCore.EntityFramework {
+namespace Cortside.AspNetCore.EntityFramework.Searches {
     public interface ISearch {
         int PageNumber { get; set; }
         int PageSize { get; set; }
