+++ src/Cortside.AspNetCore.EntityFramework/Searches/ISearchBuilder.cs
@@ -1,6 +1,6 @@
 using System.Linq;
 
-namespace Cortside.AspNetCore.EntityFramework {
+namespace Cortside.AspNetCore.EntityFramework.Searches {
     public interface ISearchBuilder<T> {
         IQueryable<T> Build(IQueryable<T> list);
     }
