+++ src/Cortside.AspNetCore.EntityFramework/Searches/Search.cs
@@ -1,4 +1,4 @@
-﻿namespace Cortside.AspNetCore.EntityFramework {
+﻿namespace Cortside.AspNetCore.EntityFramework.Searches {
     public class Search : ISearch {
         public int PageNumber { get; set; } = 1;
         public int PageSize { get; set; } = 30;
